Update header and picture with flex
This commit incorporates flex into some aspects
of this portfolio. Including mainly the header and
the picture.

I've been meaning to change the Instagram to be ligned
to the right, but struggled a big in alignment prior
to learning flex. I just applied what I learned, and
I'm pretty happy with it.